+++ user/jswrap_json.c
@@ -69,7 +69,7 @@ JsVar *jswrap_json_parse_internal(JsLex *lex) {
       return r;
     }
     case LEX_INT: {
-      int64_t v = stringToInt(jslGetTokenValueAsString(lex));
+      long long v = stringToInt(jslGetTokenValueAsString(lex));
       jslGetNextToken(lex);
       return jsvNewFromLongInteger(v);
     }
